refactor: 클래스 이름 변경, 스트림 for 문으로 변경
